requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
dotenv==0.9.9
psycopg==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
